Add files via upload
A printer prints labels for boxes and uses colors named with letters from a to m. The function printer error will receive a string and return the error rate of the printer as a string representing a rational whose numerator is the number of errors and the denominator the length of the control string. 
Example output:
s="aaabbbbhaijjjm"
printer_error(s) => "0/14"