Add files via upload
In this exercise I created a file were given an input number the iterator will ask the user to enter the same number of students' data (first name and last name). With their full names my code will create random ID numbers and generate an email account with the first letter of the first name and last name and the last three digits of the ID number. Emails and ID numbers will be created for each student and all of the student data will be appended to students_data dictionary.